Debug: use `pip list` over `poetry show`
`poetry show` primarily describes the lockfile, and not the
poetry-managed venv. (To convince yourself of this, try `poetry show`)
in a clean checkout of synapse, before running a `poetry install`. For
debugging we mainly want to see what the virtualenv contains, so use
`pip list` instead.

For completeness, `poetry show` does have some highlighting (falling
back to warning markers when ANSI colouring is unavailable) which
indicates when the virtualenv deviates from the lockfile. Its meaning
isn't fully clear and I can't find a documentation reference for it.